+++ Zooper.Bee/IWorkflowStep.cs
@@ -1,4 +1,0 @@
-namespace Zooper.Bee;
-
-// ReSharper disable once UnusedType.Global
-public interface IWorkflowStep;